tests: move MockCmd into test_utils.py
